Create GuildLaunch Import Sample
From Guildlaunch's description of the Basic Instructions:

You can use the XML format below to import raids for any game. The XML File documented below has 5 major parts.
1. Basic Raid Info - start time, end time, import version
2. PlayerInfos - who is in the raid
3. BossKills - the bosses that are killed or the 'events' in the raid
4. Join - when a player joined the raid
5. Leave - when a player left the raid
6. Loot - the items that dropped

All of the fields in the sample XML are required. However, not all of them have to have values. Things to note: The times are in unix Epoch time. Convertors are readily available on the internet via a google search.

Apologies if this is already something known or being worked on.